feat: projects from db

+++ database/migrations/000001_create_project_table.up.sql
@@ -2,9 +2,9 @@ CREATE TABLE IF NOT EXISTS project(
     id serial primary key,
     name varchar not null,
     -- in mCPU
-    cpu integer not null,
-    -- in bytes
-    memory integer not null,
-    -- in bytes
-    storage integer not null
+    cpu bigint not null,
+    -- in MiB
+    memory bigint not null,
+    -- in MiB
+    storage bigint not null
 )
